chore: renames datepicker directory to datePicker
