+++ core/components/organisms/datePicker/__stories__/index.story.tsx
@@ -1,14 +1,18 @@
 import * as React from 'react';
 import { boolean, select, date, number } from '@storybook/addon-knobs';
-import DatePicker from '@/components/organisms/datepicker';
-import InputMask from '@/components/molecules/inputMask';
+import { DatePicker, Card, InputMask } from '@/index';
 import { action } from '@storybook/addon-actions';
 
 // CSF format story
-export const withInput = () => {
-  const dateValue = date(
-    'date',
-    undefined
+export const all = () => {
+  const withInput = boolean(
+    'withInput',
+    false
+  );
+
+  const open = boolean(
+    'open',
+    false
   );
 
   const inputFormat = select(
@@ -20,6 +24,11 @@ export const withInput = () => {
   const outputFormat = select(
     'outputFormat',
     ['mm/dd/yyyy', 'dd/mm/yyyy', 'yyyy-mm-dd', 'mm-dd-yyyy', 'dd-mm-yyyy', 'yyyy-mm-dd'],
+    undefined
+  );
+
+  const dateValue = date(
+    'date',
     undefined
   );
 
@@ -66,23 +75,46 @@ export const withInput = () => {
   if (yearNav !== -1) attr.yearNav = yearNav;
   if (monthNav !== -1) attr.monthNav = monthNav;
 
+  if (withInput) {
+    return (
+      <DatePicker
+        withInput={withInput}
+        open={open}
+        inputFormat={inputFormat}
+        outputFormat={outputFormat}
+        jumpView={jumpView}
+        date={dateValue}
+        onDateChange={(currDate?: Date) => action(`on date change : ${currDate}`)()}
+        view={view}
+        firstDayOfWeek={firstDayOfWeek}
+        {...attr}
+      />
+    );
+  }
   return (
-    <DatePicker
-      withInput={true}
-      inputFormat={inputFormat}
-      outputFormat={outputFormat}
-      jumpView={jumpView}
-      date={dateValue}
-      onDateChange={(currDate?: Date, currValue?: string) => action(`on date change : ${currDate} --- ${currValue}`)()}
-      view={view}
-      firstDayOfWeek={firstDayOfWeek}
-      {...attr}
-    />
+    <Card
+      shadow="light"
+      style={{
+        maxWidth: '330px'
+      }}
+    >
+      <DatePicker
+        withInput={withInput}
+        inputFormat={inputFormat}
+        outputFormat={outputFormat}
+        jumpView={jumpView}
+        date={dateValue}
+        onDateChange={(currDate?: Date) => action(`on date change : ${currDate}`)()}
+        view={view}
+        firstDayOfWeek={firstDayOfWeek}
+        {...attr}
+      />
+    </Card>
   );
 };
 
 export default {
-  title: 'Organisms|Datepicker/Variants',
+  title: 'Organisms|DatePicker',
   component: DatePicker,
   subcomponents: { InputMask }
 };
